feat: add auth lazy loading
