+++ youtube-client-app/src/app/auth/auth-router.module.ts
@@ -1,18 +1,16 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { authGuard } from '../core/guards/auth/auth.guard';
 import { AuthPageComponent } from './pages/auth-page/auth-page.component';
 
-const routes: Routes = [
+const authRoutes: Routes = [
   {
     path: '',
     component: AuthPageComponent,
-    canActivate: [authGuard],
   },
 ];
 
 @NgModule({
-  imports: [RouterModule.forChild(routes)],
+  imports: [RouterModule.forChild(authRoutes)],
   exports: [RouterModule],
 })
 export class AuthRouterModule {}
