+++ youtube-client-app/src/app/auth/pages/auth-page/auth-page.component.ts
@@ -17,7 +17,7 @@ export class AuthPageComponent {
   constructor(
     private fb: FormBuilder,
     private authService: AuthService,
-    public router: Router,
+    public router: Router
   ) {}
 
   onSubmit() {
