+++ youtube-client-app/src/app/core/core.module.ts
@@ -3,7 +3,7 @@ import { HeaderComponent } from './components/header/header.component';
 import { SearchBarComponent } from './components/search-bar/search-bar.component';
 import { LayoutComponent } from './pages/layout/layout.component';
 import { SortComponent } from './components/sort/sort.component';
-import { AppRouterModule } from '../router/app-router.module';
+import { AppRouterModule } from '../app-router.module';
 import { SharedModule } from '../shared/shared.module';
 
 @NgModule({
