+++ youtube-client-app/src/app/youtube/components/card/card.component.html
@@ -37,6 +37,6 @@
     </div>
   </div>
   <div class="p-6 pt-0">
-    <a [routerLink]="['/details', card.id]"><app-button>More</app-button></a>
+    <a [routerLink]="['./details', card.id]"><app-button>More</app-button></a>
   </div>
 </div>
