+++ youtube-client-app/src/app/youtube/pages/details/details-page.component.html
@@ -51,12 +51,12 @@
     </div>
     <a routerLink="/"
       ><button
-        class="absolute left-[-32px] top-2 w-8 h-8 bg-red-500 rounded-l-md hover:shadow-lg hover:shadow-red-500/40"
+        class="absolute left-[-42px] w-10 h-10 bg-transparent rounded-l-md shadow-lg hover:shadow-red-500/40"
       >
         <img
           src="assets/arrow-sm-left.svg"
           alt="arrow"
-          class="w-7 h-7"
+          class="w-10 h-10"
         /></button
     ></a>
   </div>
