+++ youtube-client-app/src/app/youtube/youtube-router.module.ts
@@ -2,12 +2,17 @@ import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { YouTubePageComponent } from '../youtube/pages/youtube/youtube-page.component';
 import { authGuard } from '../core/guards/auth/auth.guard';
+import { DetailsPageComponent } from './pages/details/details-page.component';
 
 const routes: Routes = [
   {
     path: '',
     component: YouTubePageComponent,
     canActivate: [authGuard],
+  },
+  {
+    path: 'details/:id',
+    component: DetailsPageComponent,
   },
 ];
